Extract orchestrator scope-lease mixin to keep orchestrator under 1k (#222)
Move acquire/park/wake/release and satisfaction short-circuit into
orchestrator_scope_lease.py so the DAG scheduler file stays under the
1000-line maintainability bar.
